requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
langchain==0.2.16
langchain-openai==0.1.25
pytesseract==0.3.13
pdf2image==1.17.0
python-docx==1.1.2
python-magic==0.4.27
Pillow==10.4.0